Add support for $error and fix bug in ternary derivatives
$error is like $probe but should cause the simulator to exit with a
fatal error.  It had not previously been implemented.  It is basically
a direct copy of the back-end code for $strobe, with UserInfo replaced
by UserFatal.

There was also a bug in the output of ternary operator derivatives if
both the value-when-true and value-when-false (in the derivative) were
identically zero.  The ternary derivative should then have been
identically zero, but instead it was incorrectly being emitted
as (condition)?0.0:0.0.

This means that any expression in which a zero-derivative ternary was
a subexpression was not getting properly reduced, and this led to some
weird issues in the BSIM-CMG 111.  Now, if both value-if-true and
value-if-false of the derivative are identically zero, then the
derivative of the ternary is returned as "0.0".  This allows
downstream processing to emit more efficient code (like a-b gets
emitted as "(-b)" if a is 0.0, or a*b is 0.0 if either a or b is 0.0).