Make BSIM-CMG devconmap code inserted by patch
All three versions of BSIM-CMG had been manually edited in C++ form
after code generation to insert correct devConMap entries so that
connectivity checks worked properly in the presence of the temperature
node.

That meant that any regeneration of the BSIM-CMG models from Verilog
source would clobber the manually inserted code.

This commit adds a patch step to the devices' makefiles so that it
happens automatically every time the code is regenerated.  This is the
same process already being done in a couple of other devices such as
the VBIC 1.3 model.