Update VBIC patch files
The patch files in this directory are slightly outdated and in some
cases can fail to patch cleanly on some systems.  Eric discovered that
today, and when *I* try it on the public repo it worked fine, but
patch stated that there were hundreds of lines of "fuzz" required for
some of the bits (meaning the context was spotted and the patch
applied at a very different line than the one indicated in the diff).

This commit updates the patch files for VBIC so they apply without
excessive fuzz.  Further, I spotted that in one case the patch file
was looking for context that differed in case from what is currently
generated by ADMS --- a lower case "q" instead of upper case.  Somehow
on my runs this passed and the patch applied cleanly, but in Eric's
case it failed.