Update ADMS generation/comparison/install scripts
BSIM-CMG 111 was added, but these scripts weren't updated yet.  I went
ahead and patched the scripts because it was easy enough.

"generate_ADMS.sh" should be run from the top level of the Xyce source
tree, and will regenerate all ADMS devices in their own directories,
then compare the newly generated code to the stuff in
src/DeviceModelPKG.

"compare_ADMS.sh" simply repeats the compare done above and should be
run from the top level of the source tree.

"install_ADMS.sh" takes all the source code out of the directory in
which it's generated and copies it to src/DeviceModelPKG.

generate_ADMS.sh should be run every time there's a backend change and
the diffs checked to make sure no unexpected changes are happening in
the source code.   If you run it now, you will find that there are a
bunch of changes in devices due to the fix of ternary operator
derivatives made a few months ago.  In the cases that changed, the
change is inconsequential (but makes for cleaner code that the
compiler doesn't have to optimize away).