Fix some illegal instruction errors
On FreeBSD 13.4 with clang 18.1.6, parallel Xyce has started crashing
on every single netlist in the test suite.  A previously built binary
from FreeBSD 13.3 (with Clang 17.something) still runs fine.

I have tracked SOME of these failures down.

In one case it's due to a strange use in N_PDS_MPI.C of the
std::string constructor being called with two char pointer arguments,

      dest[i] = std::string(&receive[receive_displacement[i]], &receive[receive_displacement[i] + receive_count[i]]);

presumably intended to mean "copy the string from this pointer up to
that one", but this usage is not apparently part of the standard and I
can find no documentation that it is legal.   Replacing one such
instance of this usage with

      dest[i] = std::string(&receive[receive_displacement[i]], receive_count[i]);

fixes a very large number of failures.  But many more remain.

In N_TOP_ParLSUtil.C there are a few instances of using &(vector[0])
to get at the storage for a vector, but no test that the vector isn't
empty is performed first.  This leads to undefined behavior, which in
my case means "worked before, doesn't work now."

Fixing one set of these (using vector.data() instead of &(vector[0]))
fixes 56 more tests for me, but hundreds remain.