Fix another bad memory access
This commit fixes a single instance of &(vector[0]) being used without
checking that vector is not empty in N_PDS_MPI.h.

This single line was responsible for 259 test failures on FreeBSD 13.4
with Clang 18.

After fixing it, only 162 tests out of 2083 tests still fail.   This
is down from over 1900 failures before I started this exercise.