Fix more bad memory accesses
In order to make Xyce_Regression/Netlists/HB/RC/lcosc.cir run on
FreeBSD without crashing, I had to change 9 instances of &(vector[0])
to vector.data().

After fixing these 9 instances and getting that one netlist to run, a
complete run of the parallel test suite now fails only 139 tests,
meaning it fixed 23 failures in total.

I'm now getting down to code that is only invoked on small numbers of
specific test types, so I expect that it might take a few more passes
to finish ---  I don't think I'm going to find the one line that fixes
all the rest of the problems.

I am also confident that the problems that are getting fixed in this
process are the tip of the iceberg, and that many more exist that
simply aren't getting executed by the test suite.  I will not be
rooting those out.