Fix still more data access problems
This commit fixes two lines in N_ANP_SweepParam.C that were causing 31
test failures for me on FreeBSD 13.4 with Clang 18.   Both were
instances of the unconditional access of &(vector[0]) without checking
that vector isn't empty, and the crashes were because it *was* empty.

To prevent implementation-dependent failures like this and keep the
code simple, the fix is to use vector.data() (not introduced until
C++11).  That's still not safe if in fact the code downstream tries to
dereference the pointer without checking that the vector isn't empty,
because the pointer returned could be null if it is empty.

So far I have not found any instances of *that* unsafe usage.