Fix incorrect string constructor use
std::string(char *, char *) is not a documented string constructor,
and using it causes Xyce to crash with an illegal instruction trap on
FreeBSD 13.4 with Clang 18.

Replacing
   newstring=std::string(&chararray[offset],chararray[offset+count]);
with
   newstring=std::string(&chararray[offset],count);

fixes the crashes and uses a documented form of the string
constructor.

This single line fixes 46 parallel test failures, leaving only 5 to be
investigated.