Fix uninitialized variable return
All of the getMutableValue and getImmutableValue functions fail to
initialize the variable they will use to return a value.  Only in the
one case where the function is unable to parse the value of the
parameter given to it into the form requested will it return this
uninitialized value.

Clang 19 and 20, when building with full optimization, will cause Xyce
to drop core when this uninitialized value is returned.

This only happens in one of the test cases in the test suite, but
would also happen if a user had netlisting errors that trip the error
message from these functions.

This commit fixes getMutableValue<double> by initializing the "val"
variable to zero.

No other versions of the function are fixed here, this commit only
fixes the one failing test.  Clearly, the other function versions are
not ever tested for their error condition.