add prelinary option to sort sensitivities
This is in support of:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/1010
"Add option for sensitivity output to be sorted by magnitude"

For some use cases it is convenient for the sensitivities to be sorted
by magnitued.  This commit performs the first implementation of such a
capability.  It only does it for output to stdout at the moment.

It can do a separate sorting every time the output function is
called.   For a transient, the sorting could have a different
order at different time steps, so this isn't appropriate for plot files,
which have a fixed order through the whole transient.

It can also perform a single sorted output at the end of the transient,
with respect to the max magnitudes of all the sensitivities.

For transient, this is only set up for direct sensitivities.

For transient adjoint sensitivities, it would be best to set up an
"event based objective" for maximum magnitudes first.  Otherwise, the
only way for transient adjoint to produce and equivalent output at the
moment is to perform many backward integrations, which defeats that
point of using adjoints in terms of performance.

To get un-sorted output at every time point, stdoutput=1 should be added
to the .options sensitivity line in the netlist.

To change the point output to be sorted, also add sortpointoutput=1.

To get un-sorted maximum magnitude output to stdout at the end of a
transient, add maxstdoutput=1 to the .options sensitivity line.

To change the maximum magnitude output to be sorted, add sortmaxoutput=1
to that same line.